feat: add soft deletes

+++ app/Http/Controllers/ScheduleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\DTOs\ScheduleDTO;
 use App\Models\Schedule;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class ScheduleController extends Controller
@@ -21,12 +22,39 @@ class ScheduleController extends Controller
     }
 
     /**
+     * Display a listing of the resource remove soft.
+     */
+    public function softList()
+    {
+        //
+        $response = Gate::inspect('softList', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $deletedSchedules = Schedule::onlyTrashed()->get();
+        $deletedSchedulesDTO = ScheduleDTO::fromCollection($deletedSchedules);
+        return response()->json($deletedSchedulesDTO, Response::HTTP_OK);
+    }
+
+    /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
         //
         // 'day'=>'required|string|in:' . implode(',', Schedule::DAYS),
+        $response = Gate::inspect('store', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         Schedule::create($request->validated_data);
 
         return response()->json(["message" => "Schedule created successfully"], Response::HTTP_CREATED);
@@ -49,21 +77,92 @@ class ScheduleController extends Controller
     public function update(Request $request, Schedule $schedule)
     {
         //
+        $response = Gate::inspect('update', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         $schedule->update($request->validated_data);
 
         return response()->json(["message" => "Schedule has been successfully updated"], Response::HTTP_ACCEPTED);
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove soft the specified resource from storage.
      */
-    public function destroy(Schedule $schedule)
+    public function softDestroy(Schedule $schedule)
     {
         //
-        // $course->schedules()->find($schedule->id)->delete();
+        $response = Gate::inspect('softDestroy', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
 
         $schedule->delete();
+        return response()->json([
+            "message" => "the schedule has been successfully deleted"
+        ], Response::HTTP_ACCEPTED);
+    }
 
-        return response()->json(["message" => "Schedule has been successfully deleted"], Response::HTTP_ACCEPTED);
+    /**
+     * Restore the specified resource from storage.
+     */
+    public function restore($id)
+    {
+        //
+        $response = Gate::inspect('restore', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $schedule = Schedule::onlyTrashed()->find($id);
+
+        if (is_null($schedule)) {
+            return response()->json([
+                "message" => "the schedule does not exist"
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $schedule->restore();
+        return response()->json([
+            "message" => "the schedule has been successfully restored"
+        ], Response::HTTP_OK);
+    }
+
+    /**
+     * Remove permanently the specified resource from storage.
+     */
+    public function destroy($id)
+    {
+        //
+        $response = Gate::inspect('destroy', Schedule::class);
+
+        if (!$response->allowed()) {
+            return response()->json([
+                "message" => $response->message()
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $schedule = Schedule::onlyTrashed()->find($id);
+
+        if (is_null($schedule)) {
+            return response()->json([
+                "message" => "the schedule does not exist"
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $schedule->forceDelete();
+        return response()->json([
+            "message" => "the scheduler has been successfully deleted permanently"
+        ], Response::HTTP_ACCEPTED);
     }
 }
